old/pyproject.toml
[project]
name = "image-bucketing"
version = "0.1.0"
description = "Image bucketing and search using CLIP-style embeddings"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=2.2.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

